Don't cleanup if we can't tell which formula is for keeps
If the formula was updated and not yet installed, then cleanup would delete all kegs. Which is most likely not what the user wanted.

Ideally we'd do a version check to figure out which keg is newest. But we don't yet have a function that can determine with accuracy version order.