Remove slow tab completion for `brew tap`
This requires hitting the network, which makes the first invocation
slow. The results are inaccurate as it only pulls down the first page of
results. It is also prone to false-positives.

A better implementation is welcome, but in the years since I wrote this
code I can't think of a single time where I found it useful.