AprRequirement should not generate a build-time dependency
We want the apr requirement to act as a build-time requirement in case
it is satisfied (that is, the CLT is installed), as the resulting
binaries will link to the system libapr which is always present.

When it is *not* satisfied by the CLT, and we need to install the
formula, we have to treat it as a runtime dependency since the resulting
binaries will link to it.

Fixes Homebrew/homebrew#36301.
Fixes Homebrew/homebrew#36438.
Closes Homebrew/homebrew#36443.