Add Homebrew Cask completions
This completion was originally sourced from the oh-my-zsh brew cask plugin.
(https://github.com/robbyrussell/oh-my-zsh/blob/master/plugins/brew-cask/brew-cask.plugin.zsh)

The rationale for splitting this from the original omz plugin is to address the
following issues:
1. it is part of oh-my-zsh which some have chosen not to use (e.g. zprezto users)
2. it acts as a script overriding the existing brew completion rather than as a
standard autoloadable file, this means it must be manually installed rather than
automatically installed as part of compinit, and may cause issues with future
commands added to homebrew.

The changes from the plugin are to fix item 2, as well as update the commands to
those current in homebrew cask.