keg: make guile's site-dir persistent
Things that install Guile scheme objects should be doing so in
HOMEBREW_PREFIX/share/guile/site, which should be persistent rather
than a symlink to any formulae's Cellar, not even guile's necessarily as
https://github.com/Homebrew/homebrew-core/commit/f061d864d99da4064a763ba51a37ecf3cff8943c
managed to expose.

I believe guile isn't actually configured correctly to expect
this directory, but since absolutely nobody has complained to
date as far as I can find & `gnutls` seems to be the only formula
using it this way the temptation is to leave it misconfigured
& use this directory for site schemes.

I don't really know the deeper issues around messing with Guile's
configuration so going for a minimal fix seems the safest option.