brew create: add meson support
meson is quickly gaining popularity as build system, in combination with
ninja. Several Gnome projects for example are currently transitioning
from autotools to meson, mostly because it allows for Visual Studio
builds, which is impossible to accomplish with autotools.

In order to facilitate generating meson based Formulas, I added support
for meson to brew-create.